s390x: prepare device memory address space
Let's allocate the device memory information and setup the device
memory address space. The RAM size returned via SCLP is not modified. Guest
OSs which support memory devices (like virtio-mem) are expected to
consult diag500(4).

As the interaction between memory devices like virtio-mem and storage
attributes is still a bit unclear, especially when it comes to migration
but also to handling when discarding memory, let's disable CMM similarly
as we do with huge pages.

Signed-off-by: David Hildenbrand <david@redhat.com>