virtio-mem: disable reading of unplugged memory on file-based memory backends
For now, the primary focus of virtio-mem was anonymous memory only. As we
have a shared zeropage, we are able to support reading unplugged
memory, just as the spec requires. As we also want to support file-backed
memory in the future, where we don't have a shared zeropage, we have to
tell the guest to not assume it can read unplugged memory in this case.
Otherwise, even sane guests might trigger a higher memory consumption
than expected.

We might have ways in the future to still allow for reading unplugged
memory with file-backed memory backends; however, new behavior will have to
be unlocked via additional device properties.

Expose the feature VIRTIO_MEM_F_UNPLUGGED_UNREADABLE in case we have
file-backed memory and fail feature probing in case the device is not
prepared to handle this flag.

Existing guests will have to be updated to handle this feature flag.

Signed-off-by: David Hildenbrand <david@redhat.com>